pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "scheduling-lezat-backend"
version = "0.1.0"
description = "Backend API for Lezat scheduling"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "pydantic-settings>=2.3.0,<3.0.0",
  "email-validator>=2.2.0,<3.0.0",
  "pymongo>=4.8.0,<5.0.0",
  "google-genai>=1.0.0,<2.0.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.0,<1.0.0",
  "pytest>=8.2.0,<9.0.0",
  "pytest-asyncio>=0.23.0,<1.0.0",
  "ruff>=0.5.0,<1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
minversion = "7.4"
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
